+++ src/store/index.js
@@ -1,13 +1,13 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
 import config from './config';
-import * as PlvUtil from '@/utils';
+import * as PolyvUtil from '@/utils';
 
 Vue.use(Vuex);
 
 export default new Vuex.Store({
   state: {
-    isMobile: PlvUtil.isMobile(),
+    isMobile: PolyvUtil.isMobile(),
   },
   modules: {
     config
